refactor: reorganize Slate plugins and flatten directory structure
- Remove `editor/plugins/` subfolder, move all Slate plugins directly to
  `editor/` with consistent `slate-plugin.*.ts` naming convention
- Remove `history/` folder, relocate files:
  - Slate plugins → `editor/slate-plugin.{history,redoing,undoing,without-history}.ts`
  - Behavior operations → `operations/behavior.operation.history.{undo,redo}.ts`
  - `transform-operation.ts` → `internal-utils/`
  - `undo-step.ts` → `editor/`
- Move test files to dedicated folders:
  - Gherkin feature files → `gherkin-spec/`
  - Gherkin test runners → `gherkin-tests/`
  - Unit/integration tests → `tests/`
- Change plugin signatures from factory pattern `createWithX(opts) => (editor) => editor`
  to direct functions `withX(opts, editor) => editor`
- Rename `with-plugins.ts` → `slate-plugins.ts`

This consolidates related code, establishes consistent naming conventions,
and simplifies the module structure.
