+++ packages/editor/src/editor/create-editor.ts
@@ -14,13 +14,13 @@ import {createEditorPriority} from '../priority/priority.types'
 import type {EditableAPI} from '../types/editor'
 import type {PortableTextSlateEditor} from '../types/slate-editor'
 import {defaultKeyGenerator} from '../utils/key-generator'
+import {createEditableAPI} from './create-editable-api'
 import {createSlateEditor, type SlateEditor} from './create-slate-editor'
 import {createEditorDom} from './editor-dom'
 import type {EditorActor} from './editor-machine'
 import {editorMachine, rerouteExternalBehaviorEvent} from './editor-machine'
 import {getEditorSnapshot} from './editor-selector'
 import {mutationMachine, type MutationActor} from './mutation-machine'
-import {createEditableAPI} from './plugins/createWithEditableAPI'
 import {relayMachine, type RelayActor} from './relay-machine'
 import {syncMachine, type SyncActor} from './sync-machine'
