+++ packages/editor/src/editor/slate-plugin.update-selection.ts
@@ -1,8 +1,8 @@
-import {slateRangeToSelection} from '../../internal-utils/slate-utils'
-import type {PortableTextSlateEditor} from '../../types/slate-editor'
-import type {EditorActor} from '../editor-machine'
+import {slateRangeToSelection} from '../internal-utils/slate-utils'
+import type {PortableTextSlateEditor} from '../types/slate-editor'
+import type {EditorActor} from './editor-machine'
 
-export function pluginUpdateSelection({
+export function withUpdateSelection({
   editor,
   editorActor,
 }: {
