+++ packages/editor/src/editor/slate-plugin.without-history.ts
